BCELoss + Second Model + Main.py
Original FNN has a lot of changes made. Primary change made was the switch from CrossEntropyLoss to BCELoss. This is better for Binary Classification and uses the sigmoid function making it more usefull for binary classification. CrossEntropyLoss uses Softmax making it better for multi classification. For better understanding on these two, please refer to pytorch documentation.

2 files were addedmodel_test.py, this uses BCELoss and normalization on the data during preprocessing. Different from the prior model. It also has some other changes that can be looked at.

Main.py uses the last patient in the CSV file and runs it through both saved eval model states. It then produces a result for both of them.